Add support for passing functions to runEvent.
If the final argument is a function, every time a status is run, after it gets a result, it calls that function with the last return value and the current return value as arguments, and saves the result in the last return value variable. After the event has been run, it returns the final result.